Fix the placement for get_cache_block_size
